COMPILATION: _improve_ snprintf() definition guards
On windows snprintf() was explicitly implemented. That breaks
cross compile with mingw as snprintf() is already defined.

So change _WIN32 to _MSC_VER so it triggers for MSVC only. I
believe later versions of MSVC also included snprintf() so
this may still result in issues. Better add a check for specific
MSVC versions - or check if one can detect the presence of the
snprintf() symbol itself.